+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # WIP
 
-- macOS: reset fDisplayLinkRunning when releasing CVDisplayLink on hide #305 via @bumfo
+- macOS: Added `WindowMac::setPressAndHoldEnabled`, `resetPressAndHoldEnabled`, `isPressAndHoldEnabled`, `isPressAndHoldEnabledGlobally` #303 via @Folcon
+- macOS: reset `fDisplayLinkRunning` when releasing `CVDisplayLink` on hide #305 via @bumfo
 - Added Windows ARM build #306 #307 via @chirontt
 
 # 0.4.24 - Feb 24, 2026